// frontend/src/api/types.gen.ts
export type Company = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    address?: string;
    category: string;
    email?: string;
    folderName: string;
    id: string;
    legalName?: string;
    phone?: string;
    postalCode?: string;
    requiredFiles: Array<FileInfo> | null;
    shortName?: string;
    tags?: Array<string> | null;
    website?: string;
};
// ...
export type FileInfo = {
    isDirectory: boolean;
    modifiedTime: string;
    size: number;
    standardPath: string;
    targetPath: string;
};